# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Students Week 2026 food ordering API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Vendor {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  maxPlates: number;
  orderCount: number;
  createdAt: string;
}

export interface VendorInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  imageUrl?: string;
  isActive?: boolean;
  /** @minimum 0 */
  maxPlates?: number;
}

export interface VendorUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
  /** @minimum 0 */
  maxPlates?: number;
}

export interface MainDishType {
  id: number;
  mainDishId: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
}

export interface SideItem {
  id: number;
  mainDishTypeId: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  isComplementary: boolean;
  isAvailable: boolean;
}

export interface ProteinItem {
  id: number;
  mainDishTypeId: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
}

export interface DishConfig {
  id: number;
  mainDishTypeId: number;
  maxSides: number;
  maxProteins: number;
}

export interface MainDishTypeFull {
  id: number;
  mainDishId: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
  sides: SideItem[];
  proteins: ProteinItem[];
  config: DishConfig;
}

export interface MainDishTypeInput {
  mainDishId: number;
  /** @minLength 1 */
  name: string;
  imageUrl?: string;
  isAvailable?: boolean;
}

export interface MainDishTypeUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable?: boolean;
}

export interface SideItemInput {
  mainDishTypeId: number;
  /** @minLength 1 */
  name: string;
  imageUrl?: string;
  isComplementary?: boolean;
  isAvailable?: boolean;
}

export interface SideItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  isComplementary?: boolean;
  isAvailable?: boolean;
}

export interface ProteinItemInput {
  mainDishTypeId: number;
  /** @minLength 1 */
  name: string;
  imageUrl?: string;
  isAvailable?: boolean;
}

export interface ProteinItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable?: boolean;
}

export interface DishConfigInput {
  /** @minimum 0 */
  maxSides?: number;
  /** @minimum 0 */
  maxProteins?: number;
}

export interface MainDishFull {
  id: number;
  vendorId: number;
  name: string;
  createdAt: string;
  types: MainDishTypeFull[];
}

export interface MainDishInput {
  vendorId: number;
  /** @minLength 1 */
  name: string;
}

export interface MainDishUpdate {
  /** @minLength 1 */
  name?: string;
}

export interface OrderSideItem {
  id: number;
  sideId: number;
  sideName: string;
  isComplementary: boolean;
}

export interface OrderProteinItem {
  id: number;
  proteinId: number;
  proteinName: string;
}

export type OrderFullStatus = typeof OrderFullStatus[keyof typeof OrderFullStatus];


export const OrderFullStatus = {
  pending: 'pending',
  served: 'served',
} as const;

export interface OrderFull {
  id: number;
  customerName: string;
  tableNumber: string;
  vendorId: number;
  vendorName: string;
  /** @nullable */
  mainDishTypeId?: number | null;
  /** @nullable */
  mainDishTypeName?: string | null;
  /** @nullable */
  mainDishName?: string | null;
  status: OrderFullStatus;
  createdAt: string;
  sides: OrderSideItem[];
  proteins: OrderProteinItem[];
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  tableNumber: string;
  vendorId: number;
  mainDishTypeId: number;
  sideIds: number[];
  proteinIds: number[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  served: 'served',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface VendorOrderCount {
  vendorId: number;
  vendorName: string;
  count: number;
}

export interface OrderSummary {
  total: number;
  pending: number;
  served: number;
  byVendor: VendorOrderCount[];
}

export interface TableOrderGroup {
  tableNumber: string;
  totalOrders: number;
  pendingCount: number;
  servedCount: number;
  orders: OrderFull[];
}

export interface EventTable {
  id: number;
  tableNumber: string;
  isActive: boolean;
  createdAt: string;
}

export interface TableInput {
  /** @minLength 1 */
  tableNumber: string;
  isActive?: boolean;
}

export interface TableUpdate {
  /** @minLength 1 */
  tableNumber?: string;
  isActive?: boolean;
}

export interface AppConfig {
  id: number;
  maxSides: number;
  maxProteins: number;
  allowMultipleMains: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AppConfigInput {
  /** @minimum 1 */
  maxSides?: number;
  /** @minimum 1 */
  maxProteins?: number;
  allowMultipleMains?: boolean;
}

export type GetVendorsParams = {
activeOnly?: boolean;
};

export type GetMainDishesParams = {
vendorId?: number;
};

export type GetOrdersParams = {
status?: GetOrdersStatus;
vendorId?: number;
/**
 * Search by customer name or table number
 */
search?: string;
};

export type GetOrdersStatus = typeof GetOrdersStatus[keyof typeof GetOrdersStatus];


export const GetOrdersStatus = {
  pending: 'pending',
  served: 'served',
} as const;

export type GetOrderHistoryParams = {
status?: GetOrderHistoryStatus;
vendorId?: number;
search?: string;
dateFrom?: string;
dateTo?: string;
};

export type GetOrderHistoryStatus = typeof GetOrderHistoryStatus[keyof typeof GetOrderHistoryStatus];


export const GetOrderHistoryStatus = {
  pending: 'pending',
  served: 'served',
} as const;

export type GetTablesParams = {
activeOnly?: boolean;
search?: string;
};