# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Students Week 2026 food ordering API
servers:
  - url: /api
...
paths:
...
  /vendors:
    get:
      operationId: getVendors
      tags: [vendors]
      summary: List vendors
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    patch:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, served]
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Search by customer name or table number
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderFull"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderFull"
        "409":
          description: Duplicate order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/history:
    get:
      operationId: getOrderHistory
      tags: [orders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, served]
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Order history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderFull"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderFull"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderFull"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Vendor:
      type: object
      required: [id, name, isActive, maxPlates, orderCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        maxPlates:
          type: integer
        orderCount:
          type: integer
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
        maxPlates:
          type: integer
          minimum: 0
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        maxPlates:
          type: integer
          minimum: 0
...
    OrderSideItem:
      type: object
      required: [id, sideId, sideName, isComplementary]
      properties:
        id:
          type: integer
        sideId:
          type: integer
        sideName:
          type: string
        isComplementary:
          type: boolean
    OrderProteinItem:
      type: object
      required: [id, proteinId, proteinName]
      properties:
        id:
          type: integer
        proteinId:
          type: integer
        proteinName:
          type: string
    OrderFull:
      type: object
      required: [id, customerName, tableNumber, vendorId, vendorName, status, createdAt, sides, proteins]
      properties:
        id:
          type: integer
        customerName:
          type: string
        tableNumber:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        mainDishTypeId:
          type: ["integer", "null"]
        mainDishTypeName:
          type: ["string", "null"]
        mainDishName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, served]
        createdAt:
          type: string
        sides:
          type: array
          items:
            $ref: "#/components/schemas/OrderSideItem"
        proteins:
          type: array
          items:
            $ref: "#/components/schemas/OrderProteinItem"
    OrderInput:
      type: object
      required: [customerName, tableNumber, vendorId, mainDishTypeId, sideIds, proteinIds]
      properties:
        customerName:
          type: string
          minLength: 1
        tableNumber:
          type: string
          minLength: 1
        vendorId:
          type: integer
        mainDishTypeId:
          type: integer
        sideIds:
          type: array
          items:
            type: integer
        proteinIds:
          type: array
          items:
            type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, served]
...