# lib/api-zod/src/generated/types/vendorUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Students Week 2026 food ordering API
 * OpenAPI spec version: 0.3.0
 */

export interface VendorUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
  /** @minimum 0 */
  maxPlates?: number;
}